+++ src/main/java/com/quasiris/qsf/response/mapper/DefaultFacetKeyMapper.java
@@ -1,0 +1,10 @@
+package com.quasiris.qsf.response.mapper;
+
+public class DefaultFacetKeyMapper implements FacetKeyMapper {
+
+
+    @Override
+    public String map(String value) {
+        return value;
+    }
+}