+++ src/main/java/com/quasiris/qsf/response/mapper/FacetKeyMapper.java
@@ -1,0 +1,6 @@
+package com.quasiris.qsf.response.mapper;
+
+public interface FacetKeyMapper {
+
+    String map(String value);
+}